requirements.txt
openai==0.28.1
click==8.1.7
tabulate==0.9.0
python-dotenv==1.0.0
asyncio==3.4.3
pytest==7.4.0
pytest-asyncio==0.21.0